refactor: split ui/widgets/io_tray_panel.py (1101 lines) into 3 files
Extracted from monolithic io_tray_panel.py:
- thumbnail_canvas.py: ThumbnailCanvas widget, state colors, tooltips
- io_tray_actions.py: IOTrayActionsMixin with context menus, batch ops
- io_tray_panel.py: IOTrayPanel core (imports, layout, selection, D&D)

All 427 tests pass.
