+++ ui/widgets/io_tray_panel.py
@@ -13,354 +13,22 @@ from __future__ import annotations
 
 import logging
 import os
-import shutil
-import sys
-import glob as glob_module
 from PySide6.QtWidgets import (
     QWidget, QVBoxLayout, QHBoxLayout, QLabel, QScrollArea, QSplitter,
-    QToolTip, QPushButton, QMenu, QFileDialog, QMessageBox,
+    QPushButton, QMenu, QFileDialog, QMessageBox,
 )
-from PySide6.QtCore import Qt, Signal, QRect, QSize, QEvent, QUrl
-from PySide6.QtGui import QPainter, QColor, QImage, QMouseEvent, QAction, QDesktopServices
+from PySide6.QtCore import Qt, Signal
 
 from backend import ClipEntry, ClipState
 from backend.project import VIDEO_FILE_FILTER, is_image_file, is_video_file
 from ui.models.clip_model import ClipListModel
+from ui.widgets.thumbnail_canvas import ThumbnailCanvas, _STATE_COLORS  # noqa: F401
+from ui.widgets.io_tray_actions import IOTrayActionsMixin
 
 logger = logging.getLogger(__name__)
 
-# State → color mapping (matches brand palette)
-_STATE_COLORS: dict[ClipState, str] = {
-    ClipState.EXTRACTING: "#FF8C00",
-    ClipState.RAW: "#808070",
-    ClipState.MASKED: "#009ADA",
-    ClipState.READY: "#FFF203",
-    ClipState.COMPLETE: "#22C55E",
-    ClipState.ERROR: "#D10000",
-}
 
-
-class ThumbnailCanvas(QWidget):
-    """Custom-painted horizontal strip of clip thumbnail cards.
-
-    Each card is CARD_WIDTH wide and shows: thumbnail, clip name, state badge,
-    and frame count. Mouse clicks emit card_clicked with the ClipEntry.
-    """
-
-    card_clicked = Signal(object)  # ClipEntry (single left-click)
-    card_double_clicked = Signal(object)  # ClipEntry (double-click)
-    multi_select_toggled = Signal(object)  # ClipEntry (Ctrl+click toggle)
-    shift_select_requested = Signal(object)  # ClipEntry (Shift+click range)
-    context_menu_requested = Signal(object)  # ClipEntry (right-click)
-    folder_icon_clicked = Signal(object)  # ClipEntry (folder icon click)
-
-    CARD_WIDTH = 130
-    CARD_SPACING = 4
-    CARD_PADDING = 6
-    THUMB_W = 110
-    THUMB_H = 62
-
-    def __init__(
-        self,
-        parent=None,
-        show_manifest_tooltip: bool = False,
-        thumbnail_kind: str = "input",
-    ):
-        super().__init__(parent)
-        self._clips: list[ClipEntry] = []
-        self._model: ClipListModel | None = None
-        self._show_manifest_tooltip = show_manifest_tooltip
-        self._thumbnail_kind = thumbnail_kind
-        self._selected_names: set[str] = set()
-        self._hovered_name: str | None = None
-        self._thumb_cache: dict[str, QImage] = {}  # name → scaled thumbnail
-        self.setMouseTracking(True)
-        self.setMinimumHeight(100)
-
-    def set_clips(self, clips: list[ClipEntry], model: ClipListModel) -> None:
-        """Update the displayed clips and trigger repaint."""
-        # Invalidate thumbnail cache when clip list changes
-        new_names = {c.name for c in clips}
-        old_names = {c.name for c in self._clips}
-        if new_names != old_names:
-            self._thumb_cache = {k: v for k, v in self._thumb_cache.items()
-                                 if k in new_names}
-        self._clips = list(clips)
-        self._model = model
-        total_w = max(1, len(clips) * (self.CARD_WIDTH + self.CARD_SPACING))
-        self.setFixedWidth(total_w)
-        self.update()
-
-    def set_selected(self, name: str | None) -> None:
-        """Set single-selected clip (clears multi-select, draws highlight border)."""
-        new_set = {name} if name else set()
-        if self._selected_names != new_set:
-            self._selected_names = new_set
-            self.update()
-
-    def set_multi_selected(self, names: set[str]) -> None:
-        """Set the multi-selected clip names."""
-        if self._selected_names != names:
-            self._selected_names = set(names)
-            self.update()
-
-    def paintEvent(self, event) -> None:
-        if not self._clips:
-            return
-
-        p = QPainter(self)
-        p.setRenderHint(QPainter.Antialiasing, False)
-
-        for i, clip in enumerate(self._clips):
-            x = i * (self.CARD_WIDTH + self.CARD_SPACING)
-            card_rect = QRect(x, 0, self.CARD_WIDTH, self.height())
-
-            # Skip cards not in the visible region
-            if not card_rect.intersects(event.rect()):
-                continue
-
-            self._paint_card(p, card_rect, clip)
-
-        p.end()
-
-    def _paint_card(self, p: QPainter, rect: QRect, clip: ClipEntry) -> None:
-        pad = self.CARD_PADDING
-        is_selected = clip.name in self._selected_names
-        is_hovered = clip.name == self._hovered_name and not is_selected
-
-        # Card background
-        if is_selected:
-            bg = QColor("#252413")
-        elif is_hovered:
-            bg = QColor("#1E1D0A")
-        else:
-            bg = QColor("#1A1900")
-        p.fillRect(rect, bg)
-
-        # Border — yellow for selected, subtle yellow for hover, default otherwise
-        if is_selected:
-            p.setPen(QColor("#FFF203"))
-            p.drawRect(rect.adjusted(0, 0, -1, -1))
-            p.drawRect(rect.adjusted(1, 1, -2, -2))  # 2px border
-        elif is_hovered:
-            p.setPen(QColor(255, 242, 3, 100))  # subtle yellow glow
-            p.drawRect(rect.adjusted(0, 0, -1, -1))
-        else:
-            p.setPen(QColor("#2A2910"))
-            p.drawRect(rect.adjusted(0, 0, -1, -1))
-
-        # Thumbnail
-        thumb_rect = QRect(
-            rect.x() + (self.CARD_WIDTH - self.THUMB_W) // 2,
-            rect.y() + pad,
-            self.THUMB_W,
-            self.THUMB_H,
-        )
-        # Use cached scaled thumbnail to avoid expensive SmoothTransformation
-        # on every repaint.
-        scaled = self._thumb_cache.get(clip.name)
-        if scaled is None and self._model:
-            thumb = self._model.get_thumbnail(clip.name, kind=self._thumbnail_kind)
-            if isinstance(thumb, QImage) and not thumb.isNull():
-                scaled = thumb.scaled(
-                    self.THUMB_W, self.THUMB_H,
-                    Qt.KeepAspectRatio, Qt.SmoothTransformation,
-                )
-                self._thumb_cache[clip.name] = scaled
-        if scaled is not None:
-            dx = thumb_rect.x() + (self.THUMB_W - scaled.width()) // 2
-            dy = thumb_rect.y() + (self.THUMB_H - scaled.height()) // 2
-            p.drawImage(dx, dy, scaled)
-        else:
-            p.fillRect(thumb_rect, QColor("#0A0A00"))
-            p.setPen(QColor("#3A3A30"))
-            p.drawRect(thumb_rect.adjusted(0, 0, -1, -1))
-
-        # State badge (top-right over thumbnail, with background pill)
-        badge_color = QColor(_STATE_COLORS.get(clip.state, "#808070"))
-        font = p.font()
-        font.setPointSize(8)
-        font.setBold(True)
-        p.setFont(font)
-        badge_text = clip.state.value
-        metrics = p.fontMetrics()
-        text_w = metrics.horizontalAdvance(badge_text)
-        bg_rect = QRect(
-            rect.x() + self.CARD_WIDTH - pad - text_w - 6,
-            rect.y() + pad,
-            text_w + 6, 14,
-        )
-        p.fillRect(bg_rect, QColor(0, 0, 0, 128))
-        p.setPen(badge_color)
-        p.drawText(bg_rect, Qt.AlignCenter, badge_text)
-
-        # Clip name (below thumbnail, with background)
-        text_y = rect.y() + pad + self.THUMB_H + 4
-        font.setPointSize(9)
-        font.setBold(True)
-        p.setFont(font)
-        name_rect = QRect(rect.x() + pad, text_y, self.CARD_WIDTH - pad * 2, 16)
-        metrics = p.fontMetrics()
-        elided = metrics.elidedText(clip.name, Qt.ElideRight, name_rect.width())
-        p.fillRect(name_rect, QColor(0, 0, 0, 128))
-        p.setPen(QColor("#E0E0E0"))
-        p.drawText(name_rect, Qt.AlignLeft | Qt.AlignVCenter, elided)
-
-        # Source type badge (top-left over thumbnail, input cards only)
-        if not self._show_manifest_tooltip and clip.source_type != "unknown":
-            src_icon = "\U0001F39E" if clip.source_type == "video" else "\U0001F4F7"  # film frames / camera
-            src_font = p.font()
-            src_font.setPointSize(9)
-            src_font.setBold(False)
-            p.setFont(src_font)
-            src_rect = QRect(rect.x() + pad, rect.y() + pad, 18, 18)
-            p.fillRect(src_rect, QColor(0, 0, 0, 140))
-            p.setPen(QColor("#C0C0A0"))
-            p.drawText(src_rect, Qt.AlignCenter, src_icon)
-
-        # Frame count (below name, with background)
-        if clip.input_asset:
-            font.setPointSize(8)
-            font.setBold(False)
-            p.setFont(font)
-            info_rect = QRect(rect.x() + pad, text_y + 14, self.CARD_WIDTH - pad * 2, 14)
-            info_text = f"{clip.input_asset.frame_count} frames"
-            if clip.input_asset.asset_type == "video":
-                info_text += " (video)"
-            elif clip.source_type == "sequence":
-                info_text += " (imported)"
-            p.fillRect(info_rect, QColor(0, 0, 0, 128))
-            p.setPen(QColor("#808070"))
-            p.drawText(info_rect, Qt.AlignLeft | Qt.AlignVCenter, info_text)
-
-        # Folder icon (top-left of thumbnail, export cards only)
-        if self._show_manifest_tooltip:
-            icon_size = 18
-            icon_rect = QRect(rect.x() + pad, rect.y() + pad, icon_size, icon_size)
-            is_icon_hovered = (clip.name == self._hovered_name
-                               and hasattr(self, '_hover_pos')
-                               and icon_rect.contains(self._hover_pos))
-            bg_alpha = 200 if is_icon_hovered else 140
-            p.fillRect(icon_rect, QColor(0, 0, 0, bg_alpha))
-            folder_font = p.font()
-            folder_font.setPointSize(10)
-            folder_font.setBold(False)
-            p.setFont(folder_font)
-            p.setPen(QColor("#FFF203") if is_icon_hovered else QColor("#C0C0A0"))
-            p.drawText(icon_rect, Qt.AlignCenter, "\U0001F4C2")
-
-    def mouseMoveEvent(self, event: QMouseEvent) -> None:
-        clip = self._card_at(event.position().x())
-        name = clip.name if clip else None
-        self._hover_pos = event.position().toPoint()
-        if name != self._hovered_name:
-            self._hovered_name = name
-        self.update()
-
-    def leaveEvent(self, event) -> None:
-        if self._hovered_name is not None:
-            self._hovered_name = None
-            self.update()
-
-    def _folder_icon_rect(self, clip_index: int) -> QRect:
-        """Return the folder icon QRect for a given card index."""
-        x = clip_index * (self.CARD_WIDTH + self.CARD_SPACING)
-        return QRect(x + self.CARD_PADDING, self.CARD_PADDING, 18, 18)
-
-    def mousePressEvent(self, event: QMouseEvent) -> None:
-        if event.button() == Qt.LeftButton and self._clips:
-            clip = self._card_at(event.position().x())
-            if clip:
-                # Check folder icon click (export cards only)
-                if self._show_manifest_tooltip:
-                    idx = int(event.position().x() // (self.CARD_WIDTH + self.CARD_SPACING))
-                    if self._folder_icon_rect(idx).contains(event.position().toPoint()):
-                        self.folder_icon_clicked.emit(clip)
-                        return
-                from ui.sounds.audio_manager import UIAudio
-                UIAudio.click()
-                if event.modifiers() & Qt.ShiftModifier:
-                    # Shift+click: range select from anchor to this clip
-                    self.shift_select_requested.emit(clip)
-                elif event.modifiers() & Qt.ControlModifier:
-                    # Ctrl+click: toggle in/out of multi-selection
-                    self.multi_select_toggled.emit(clip)
-                else:
-                    # Plain click: single-select
-                    self.card_clicked.emit(clip)
-        elif event.button() == Qt.RightButton and self._clips:
-            clip = self._card_at(event.position().x())
-            if clip:
-                self.context_menu_requested.emit(clip)
-
-    def mouseDoubleClickEvent(self, event: QMouseEvent) -> None:
-        if event.button() == Qt.LeftButton and self._clips:
-            clip = self._card_at(event.position().x())
-            if clip:
-                self.card_double_clicked.emit(clip)
-
-    def sizeHint(self) -> QSize:
-        w = max(1, len(self._clips) * (self.CARD_WIDTH + self.CARD_SPACING))
-        return QSize(w, 100)
-
-    def _card_at(self, x: float) -> ClipEntry | None:
-        """Return the ClipEntry under the given x position, or None."""
-        if not self._clips:
-            return None
-        idx = int(x // (self.CARD_WIDTH + self.CARD_SPACING))
-        if 0 <= idx < len(self._clips):
-            return self._clips[idx]
-        return None
-
-    def event(self, ev: QEvent) -> bool:
-        if ev.type() == QEvent.ToolTip and self._show_manifest_tooltip:
-            clip = self._card_at(ev.position().x())
-            tip = _format_manifest_tooltip(clip) if clip else ""
-            if tip:
-                QToolTip.showText(ev.globalPosition().toPoint(), tip, self)
-            else:
-                QToolTip.hideText()
-            return True
-        return super().event(ev)
-
-
-def _format_manifest_tooltip(clip: ClipEntry) -> str:
-    """Build a tooltip string from the clip's .corridorkey_manifest.json."""
-    manifest = clip._read_manifest()
-    if manifest is None:
-        return ""
-
-    lines: list[str] = [f"<b>{clip.name}</b> — Export Settings"]
-
-    # Outputs + formats
-    enabled = manifest.get("enabled_outputs", [])
-    formats = manifest.get("formats", {})
-    if enabled:
-        out_parts = []
-        for name in enabled:
-            fmt = formats.get(name, "?").upper()
-            out_parts.append(f"{name.upper()} ({fmt})")
-        lines.append(f"<b>Outputs:</b> {', '.join(out_parts)}")
-
-    # Params
-    params = manifest.get("params", {})
-    if params:
-        cs = "Linear" if params.get("input_is_linear") else "sRGB"
-        lines.append(f"<b>Color Space:</b> {cs}")
-        ds = params.get("despill_strength", 1.0)
-        lines.append(f"<b>Despill:</b> {ds:.0%}")
-        rs = params.get("refiner_scale", 1.0)
-        lines.append(f"<b>Refiner:</b> {rs:.0%}")
-        if params.get("auto_despeckle"):
-            sz = params.get("despeckle_size", 400)
-            lines.append(f"<b>Despeckle:</b> On (size {sz})")
-        else:
-            lines.append(f"<b>Despeckle:</b> Off")
-
-    return "<br>".join(lines)
-
-
-class IOTrayPanel(QWidget):
+class IOTrayPanel(IOTrayActionsMixin, QWidget):
     """Bottom panel with Input and Exports thumbnail strips.
 
     Input section shows all loaded clips with + ADD button.
@@ -656,386 +324,6 @@ class IOTrayPanel(QWidget):
         """Return all clips whose names are in the selection set."""
         names = self._input_canvas._selected_names
         return [c for c in self._model.clips if c.name in names]
-
-    # ── Context menu (INPUT cards) ──
-
-    def _on_context_menu(self, clip: ClipEntry) -> None:
-        """Show right-click context menu for a clip card.
-
-        If the right-clicked clip is not in the current selection,
-        single-select it first (standard behaviour).
-        """
-        if clip.name not in self._input_canvas._selected_names:
-            self._input_canvas.set_selected(clip.name)
-            self.clip_clicked.emit(clip)
-            self.selection_changed.emit([clip])
-
-        selected = self.get_selected_clips()
-        n = len(selected)
-        multi = n > 1
-
-        menu = QMenu(self)
-
-        # Run Extraction — for clips that have a video source and need frames
-        from backend.clip_state import ClipState
-        needs_extract = [
-            c for c in selected
-            if c.state == ClipState.EXTRACTING
-            or (c.input_asset and c.input_asset.asset_type == "video")
-        ]
-        if needs_extract:
-            label_ext = (f"Run Extraction ({len(needs_extract)} clips)"
-                         if len(needs_extract) > 1 else "Run Extraction")
-            extract_action = QAction(label_ext, self)
-            extract_action.triggered.connect(
-                lambda: self.extract_requested.emit(needs_extract))
-            menu.addAction(extract_action)
-            menu.addSeparator()
-
-        # Rename — single only
-        rename_action = QAction("Rename...", self)
-        rename_action.setEnabled(not multi)
-        rename_action.triggered.connect(lambda: self._rename_clip(clip))
-        menu.addAction(rename_action)
-
-        # Open in file manager — single only
-        _fm = "Finder" if sys.platform == "darwin" else "Explorer"
-        explorer_action = QAction(f"Open in {_fm}", self)
-        explorer_action.setEnabled(not multi)
-        explorer_action.triggered.connect(lambda: self._open_in_explorer(clip))
-        menu.addAction(explorer_action)
-
-        menu.addSeparator()
-
-        # Clear Mask — only show when there's a VideoMamaMaskHint to clear
-        any_mask = any(c.mask_asset is not None for c in selected)
-        if any_mask:
-            label_mask = f"Clear Mask ({n} clips)" if multi else "Clear Mask"
-            clear_mask_action = QAction(label_mask, self)
-            clear_mask_action.triggered.connect(lambda: self._clear_mask_batch(selected))
-            menu.addAction(clear_mask_action)
-
-        # Clear Alpha — only show when there's an AlphaHint to clear
-        any_alpha = any(c.alpha_asset is not None for c in selected)
-        if any_alpha:
-            label_alpha = f"Clear Alpha ({n} clips)" if multi else "Clear Alpha"
-            clear_alpha_action = QAction(label_alpha, self)
-            clear_alpha_action.triggered.connect(lambda: self._clear_alpha_batch(selected))
-            menu.addAction(clear_alpha_action)
-
-        # Clear Outputs — only show when there are outputs to clear
-        any_outputs = any(c.has_outputs for c in selected)
-        if any_outputs:
-            label_clear = f"Clear Outputs ({n} clips)" if multi else "Clear Outputs"
-            clear_action = QAction(label_clear, self)
-            clear_action.triggered.connect(lambda: self._clear_outputs_batch(selected))
-            menu.addAction(clear_action)
-
-        # Clear All — show when there's any generated data to clear
-        if any_mask or any_alpha or any_outputs:
-            menu.addSeparator()
-            label_all = f"Clear All ({n} clips)" if multi else "Clear All"
-            clear_all_action = QAction(label_all, self)
-            clear_all_action.triggered.connect(lambda: self._clear_all_batch(selected))
-            menu.addAction(clear_all_action)
-
-        # Remove...
-        label_remove = f"Remove ({n} clips)..." if multi else "Remove..."
-        remove_action = QAction(label_remove, self)
-        remove_action.triggered.connect(lambda: self._remove_dialog(selected))
-        menu.addAction(remove_action)
-
-        from PySide6.QtGui import QCursor
-        menu.exec(QCursor.pos())
-
-    def _on_export_context_menu(self, clip: ClipEntry) -> None:
-        """Show right-click context menu for an export card."""
-        menu = QMenu(self)
-
-        # Export Video — list each available output subdirectory
-        if clip.state == ClipState.COMPLETE and hasattr(clip, 'output_dir'):
-            output_dir = clip.output_dir
-            if os.path.isdir(output_dir):
-                subdirs = sorted(
-                    d for d in os.listdir(output_dir)
-                    if os.path.isdir(os.path.join(output_dir, d))
-                    and os.listdir(os.path.join(output_dir, d))
-                )
-                if subdirs:
-                    for subdir in subdirs:
-                        src = os.path.join(output_dir, subdir)
-                        action = QAction(f"Export {subdir} as Video...", self)
-                        action.triggered.connect(
-                            lambda checked=False, c=clip, s=src: self.export_video_requested.emit(c, s)
-                        )
-                        menu.addAction(action)
-                    menu.addSeparator()
-
-        # Open containing folder (Output directory)
-        output_dir = os.path.join(clip.root_path, "Output")
-        if not os.path.isdir(output_dir):
-            output_dir = clip.root_path
-
-        open_action = QAction("Open Containing Folder", self)
-        open_action.triggered.connect(
-            lambda: QDesktopServices.openUrl(QUrl.fromLocalFile(output_dir))
-        )
-        menu.addAction(open_action)
-
-        from PySide6.QtGui import QCursor
-        menu.exec(QCursor.pos())
-
-    def _open_export_folder(self, clip: ClipEntry) -> None:
-        """Open the export/output folder for a clip."""
-        output_dir = os.path.join(clip.root_path, "Output")
-        if not os.path.isdir(output_dir):
-            output_dir = clip.root_path
-        if os.path.isdir(output_dir):
-            QDesktopServices.openUrl(QUrl.fromLocalFile(output_dir))
-
-    def _rename_clip(self, clip: ClipEntry) -> None:
-        """Prompt user to rename a clip's display name."""
-        from PySide6.QtWidgets import QInputDialog
-        from backend.project import set_display_name
-
-        current = clip.name
-        new_name, ok = QInputDialog.getText(
-            self, "Rename Clip", "New name:", text=current,
-        )
-        if not ok or not new_name.strip() or new_name.strip() == current:
-            return
-        set_display_name(clip.root_path, new_name.strip())
-        clip.find_assets()  # re-reads display_name into clip.name
-        self._rebuild()
-
-    def _open_in_explorer(self, clip: ClipEntry) -> None:
-        if os.path.isdir(clip.root_path):
-            QDesktopServices.openUrl(QUrl.fromLocalFile(clip.root_path))
-
-    def _clear_mask_batch(self, clips: list[ClipEntry]) -> None:
-        """Delete VideoMamaMaskHint folder from disk for one or more clips."""
-        names = ", ".join(c.name for c in clips[:3])
-        if len(clips) > 3:
-            names += f" (+{len(clips) - 3} more)"
-        confirm = QMessageBox.question(
-            self, "Clear Mask",
-            f"Delete tracked masks for {len(clips)} clip(s)?\n{names}\n\n"
-            "This will remove all SAM2 mask frames from disk.",
-            QMessageBox.Yes | QMessageBox.No, QMessageBox.No,
-        )
-        if confirm != QMessageBox.Yes:
-            return
-
-        for clip in clips:
-            mask_dir = os.path.join(clip.root_path, "VideoMamaMaskHint")
-            if os.path.isdir(mask_dir):
-                shutil.rmtree(mask_dir, ignore_errors=True)
-            for candidate in glob_module.glob(os.path.join(clip.root_path, "VideoMamaMaskHint.*")):
-                if os.path.isfile(candidate) and is_video_file(candidate):
-                    os.remove(candidate)
-            clip.mask_asset = None
-            clip.find_assets()
-            self._model.update_clip_state(clip.name, clip.state)
-
-        self._model.layoutChanged.emit()
-        if clips:
-            self.clip_clicked.emit(clips[0])
-        logger.info(f"Cleared masks for {len(clips)} clip(s)")
-
-    def _clear_all_batch(self, clips: list[ClipEntry]) -> None:
-        """Delete masks, alpha hints, and outputs for one or more clips."""
-        names = ", ".join(c.name for c in clips[:3])
-        if len(clips) > 3:
-            names += f" (+{len(clips) - 3} more)"
-        confirm = QMessageBox.question(
-            self, "Clear All",
-            f"Remove ALL generated data for {len(clips)} clip(s)?\n{names}\n\n"
-            "This will delete masks, alpha hints, and all output frames.",
-            QMessageBox.Yes | QMessageBox.No, QMessageBox.No,
-        )
-        if confirm != QMessageBox.Yes:
-            return
-
-        for clip in clips:
-            # Masks
-            mask_dir = os.path.join(clip.root_path, "VideoMamaMaskHint")
-            if os.path.isdir(mask_dir):
-                shutil.rmtree(mask_dir, ignore_errors=True)
-            for candidate in glob_module.glob(os.path.join(clip.root_path, "VideoMamaMaskHint.*")):
-                if os.path.isfile(candidate) and is_video_file(candidate):
-                    os.remove(candidate)
-            clip.mask_asset = None
-
-            # Alpha
-            alpha_dir = os.path.join(clip.root_path, "AlphaHint")
-            if os.path.isdir(alpha_dir):
-                shutil.rmtree(alpha_dir, ignore_errors=True)
-            for candidate in glob_module.glob(os.path.join(clip.root_path, "AlphaHint.*")):
-                if os.path.isfile(candidate) and is_video_file(candidate):
-                    os.remove(candidate)
-            clip.alpha_asset = None
-
-            # Outputs
-            output_dir = clip.output_dir
-            for subdir in ("FG", "Matte", "Comp", "Processed"):
-                d = os.path.join(output_dir, subdir)
-                if os.path.isdir(d):
-                    for f in os.listdir(d):
-                        fpath = os.path.join(d, f)
-                        if os.path.isfile(fpath):
-                            os.remove(fpath)
-            manifest = os.path.join(output_dir, ".corridorkey_manifest.json")
-            if os.path.isfile(manifest):
-                os.remove(manifest)
-
-            clip.find_assets()
-            self._model.update_clip_state(clip.name, clip.state)
-
-        self._model.layoutChanged.emit()
-        if clips:
-            self.clip_clicked.emit(clips[0])
-        logger.info(f"Cleared all generated data for {len(clips)} clip(s)")
-
-    def _clear_alpha_batch(self, clips: list[ClipEntry]) -> None:
-        """Delete AlphaHint folder from disk for one or more clips."""
-        names = ", ".join(c.name for c in clips[:3])
-        if len(clips) > 3:
-            names += f" (+{len(clips) - 3} more)"
-        confirm = QMessageBox.question(
-            self, "Clear Alpha",
-            f"Delete AlphaHint for {len(clips)} clip(s)?\n{names}\n\n"
-            "This will remove all generated alpha hint frames from disk.",
-            QMessageBox.Yes | QMessageBox.No, QMessageBox.No,
-        )
-        if confirm != QMessageBox.Yes:
-            return
-
-        for clip in clips:
-            alpha_dir = os.path.join(clip.root_path, "AlphaHint")
-            if os.path.isdir(alpha_dir):
-                shutil.rmtree(alpha_dir, ignore_errors=True)
-            for candidate in glob_module.glob(os.path.join(clip.root_path, "AlphaHint.*")):
-                if os.path.isfile(candidate) and is_video_file(candidate):
-                    os.remove(candidate)
-            clip.alpha_asset = None
-            clip.find_assets()  # re-scan disk, updates alpha_asset and state
-            self._model.update_clip_state(clip.name, clip.state)
-
-        self._model.layoutChanged.emit()
-        # Re-select first affected clip so the viewer rebuilds its FrameIndex
-        # (clears stale ALPHA button + scrubber coverage bar)
-        if clips:
-            self.clip_clicked.emit(clips[0])
-        logger.info(f"Cleared AlphaHint for {len(clips)} clip(s)")
-
-    def _clear_outputs_batch(self, clips: list[ClipEntry]) -> None:
-        """Clear output files for one or more clips."""
-        names = ", ".join(c.name for c in clips[:3])
-        if len(clips) > 3:
-            names += f" (+{len(clips) - 3} more)"
-        confirm = QMessageBox.question(
-            self, "Clear Outputs",
-            f"Remove all output files for {len(clips)} clip(s)?\n{names}\n\n"
-            "This will delete FG, Matte, Comp, and Processed frames.",
-            QMessageBox.Yes | QMessageBox.No, QMessageBox.No,
-        )
-        if confirm != QMessageBox.Yes:
-            return
-
-        total_cleared = 0
-        for clip in clips:
-            output_dir = clip.output_dir
-            cleared = 0
-            for subdir in ("FG", "Matte", "Comp", "Processed"):
-                d = os.path.join(output_dir, subdir)
-                if os.path.isdir(d):
-                    for f in os.listdir(d):
-                        fpath = os.path.join(d, f)
-                        if os.path.isfile(fpath):
-                            os.remove(fpath)
-                            cleared += 1
-            manifest = os.path.join(output_dir, ".corridorkey_manifest.json")
-            if os.path.isfile(manifest):
-                os.remove(manifest)
-
-            if clip.state == ClipState.COMPLETE:
-                clip.state = ClipState.READY
-                self._model.update_clip_state(clip.name, ClipState.READY)
-            total_cleared += cleared
-
-        self._model.layoutChanged.emit()
-        # Re-select first affected clip so the viewer rebuilds its FrameIndex
-        if clips:
-            self.clip_clicked.emit(clips[0])
-        logger.info(f"Cleared {total_cleared} output files across {len(clips)} clip(s)")
-
-    def _remove_dialog(self, clips: list[ClipEntry]) -> None:
-        """Show remove confirmation dialog with Remove from List / Delete from Disk options."""
-        n = len(clips)
-        title = f"Remove {n} clip{'s' if n > 1 else ''}?"
-
-        paths_text = "\n".join(c.root_path for c in clips[:5])
-        if n > 5:
-            paths_text += f"\n... and {n - 5} more"
-
-        msg = QMessageBox(self)
-        msg.setWindowTitle(title)
-        msg.setIcon(QMessageBox.Warning)
-        msg.setText(
-            f"How would you like to remove {n} clip{'s' if n > 1 else ''}?"
-        )
-        msg.setInformativeText(paths_text)
-
-        btn_list = msg.addButton("Remove from List", QMessageBox.AcceptRole)
-        btn_disk = msg.addButton("Delete from Disk", QMessageBox.DestructiveRole)
-        msg.addButton(QMessageBox.Cancel)
-
-        msg.exec()
-        clicked = msg.clickedButton()
-
-        if clicked == btn_list:
-            self._remove_clips_from_list(clips)
-        elif clicked == btn_disk:
-            self._delete_clips_from_disk(clips)
-
-    def _remove_clips_from_list(self, clips: list[ClipEntry]) -> None:
-        """Remove clips from the list (files stay on disk).
-
-        Records removed clip folder names in project.json so they don't
-        reappear on the next project rescan.  Uses folder_name (stable
-        on-disk identity) rather than display name which is mutable.
-        """
-        from backend.project import add_removed_clip, is_v2_project
-
-        names = {c.name for c in clips}
-
-        # Persist removal in project.json for v2 projects
-        for clip in clips:
-            # v2 clip root_path: .../project_dir/clips/clip_folder
-            parent = os.path.dirname(clip.root_path)
-            if os.path.basename(parent) == "clips":
-                project_dir = os.path.dirname(parent)
-                if is_v2_project(project_dir):
-                    add_removed_clip(project_dir, clip.folder_name)
-
-        # Remove in reverse index order to avoid index shifting
-        indices = [i for i, c in enumerate(self._model.clips) if c.name in names]
-        for i in sorted(indices, reverse=True):
-            self._model.remove_clip(i)
-        logger.info(f"Removed {len(clips)} clip(s) from list")
-
-    def _delete_clips_from_disk(self, clips: list[ClipEntry]) -> None:
-        """Delete clip project folders from disk."""
-        for clip in clips:
-            if os.path.isdir(clip.root_path):
-                shutil.rmtree(clip.root_path, ignore_errors=True)
-                logger.info(f"Deleted from disk: {clip.root_path}")
-        # Remove from model (skip persistence — folder is gone, won't reappear)
-        names = {c.name for c in clips}
-        indices = [i for i, c in enumerate(self._model.clips) if c.name in names]
-        for i in sorted(indices, reverse=True):
-            self._model.remove_clip(i)
-        logger.info(f"Deleted {len(clips)} clip(s) from disk")
 
     # ── Selection highlight ──
